# Issue 3479: Increase client throttling limit