Signed-off-by: Vijay Srinivasaraghavan <vijayaraghavan.srinivasaraghavan@emc.com>

**Change log description**  
  * increased the throttling limit by a factor of ~~10~~ 5 to ensure the clients are not blocked 

**Purpose of the change**  
To address the issue https://github.com/pravega/pravega/issues/3479

**What the code does**  
Increased the throttling configuration setting in `AppendProcessor`  

**How to verify it**  
test by writing more/frequent data with/without the fix and observe the write throughput difference