@vijikarthi 2019-03-20
> Do you observe a performance difference with this change?

Yes.