@RaulGracia 2019-03-22
@kmgowda @vijikarthi perhaps adding in the PR some numeric values/improvement percentages specifying the workload used may help reviewers to better understand the extent of this change.