@kmgowda 2019-03-26
please find the detailed results of the performance difference between the master and this pull request 
note that there major improvement in the throughput for the events sizes 1K and above
However , for 100 bytes I have not any seen any difference in terms of throughput and latencies