@vijikarthi 2019-03-26
> I'm not convinced this is a good idea for the following reasons:
> 
> * The watermarks changed here determine when to pause/resume a connection for throttling purposes. If we increase it, then we risk overwhelming the server. Increasing it by 10x possibly means that we can handle 10x fewer concurrent writers pounding the server (I'm speculating).
> * Accumulating more data might have an impact to latency. I can understand how making more data available to the server increases throughput, though.

There are pros and cons to the settings. If we keep the value to current (which looks like an arbitrary value), then we expect to run more number of writers to compensate the ingest throughput limitation. Increasing the number of writers will increase the no of TCP connections (total segments x connection) and I am not sure if that will have any impact to the system? We need to strike some balance and with the changed settings, I have seen a very decent performance improvement for the Writer.