@fpj 2019-03-27
@vijikarthi 

> If we keep the value to current (which looks like an arbitrary value), then we expect to run more number of writers to compensate the ingest throughput limitation. 

Agreed, we have not done a sensitive analysis to determine the optimal value. I do believe we need to optimize for multiple writers, which is likely to be the common case for us.

> Increasing the number of writers will increase the no of TCP connections (total segments x connection) and I am not sure if that will have any impact to the system?

The connection pooling work that is going on should address this concern at least partially.

> We need to strike some balance and with the changed settings, I have seen a very decent performance improvement for the Writer.

Right, that's possible, but it looks like we have tried a single value, could we do a bit more exploration to see what other values give us?