@vijikarthi 2019-04-05
I have tested 10KB event size for a total of 10GB records for varying writer(s) with and without the fix. Here is the throughput comparison.

| writer(s) | Throughput (with-fix) | Throughput (without-fix) |
| --- | --- | --- |
| 1 | 170 MB/s | 127 MB/s |
| 5 | 240 MB/s | 198 MB/s |
| 10 | 264 MB/s | 225 MB/s |
| 20 | 264 MB/s | 232 MB/s |