@tkaitchuck 2019-04-06
@vijikarthi
The fact that this is increasing throughput at 20 writers does indicate the current value is too small. 
What happens with other values? Like say 4x increase instead of 10x?