@kmgowda 2019-04-09
Hi Tom, 
    please see the "pravega-perf-testing" channel.. I have published the results with 1x, 2x, 4x, 6x, 8x and 10x (this PR changes).