@kmgowda 2019-04-17
Hi Tom/vijay,
   I did the testing with 10 producers with 10 segments.
There is not much improvement in case of multi producers.
The detailed results are published in the pravega-perf-testing channel.