@vijikarthi 2019-04-29
Tested with 10 producers and 10 segments for varying watermark/throttling limits ranging from 2x to 10x from the original values (HIGH_WATER_MARK  128 KB and LOW_WATER_MARK  64 KB) suggests that up to 5x, the performance is looking good and afterwards not much of improvement. The throttling limit hence is changed to 5x of the original value.  

![image](https://user-images.githubusercontent.com/5560671/56928453-b4edea00-6a8b-11e9-90a1-a15f0b5b734f.png)